Search feature complete
Made changes to the services.js file: tag filter and category filter. Also added a new filter in the services.js file to be used by the search bar. Alignment for the search bar is good, but the alignment for the Clear Search button is not ideal; cannot change alignment, because it is inherited from the parent layout.

The search bar has auto complete. Users can search by categories or tags. Added error messages to the search functionality, but it is not working as expected.